refactor: reorganize code structure and improve class definitions
