+++ graphic/config.py
@@ -1,0 +1,2 @@
+SCENE_SIZE = 5000
+FONT_SIZE = 14
